feature: messaging with firebase cont.
updated user interface.

currently missing some forms for out database that we are looking to fill. will be adding the appropriate entry fields soon

using MainActivity_ as a paralell development activity [with the underscore] to revert back to the standard app version see TODO: on SignInActivity

After messaging feautures are complete and debugged I will replace / merge the old mainActivity with the current one.